requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.22.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.6
python-dotenv==1.0.1
httpx==0.24.0
alembic==1.11.1
# pydantic must be <2 to match FastAPI 0.99.x
pydantic==1.10.19
requests==2.31.0


